gpt_database/requirements.txt
uvicorn
fastapi
pydantic
openai
sqlalchemy
python-dotenv
psycopg2-binary
mysql-connector-python